general code improvements
