+++ src/ui/components/Form/PinFormComponent.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import {
   HStack,
   Input,
@@ -10,8 +10,10 @@ import {
 import {
   ButtonDisableTheme,
   ButtonSecondaryTheme,
-} from '../../../assets/chakra/appStyle' // adjust the import path as needed
+} from '../../../assets/chakra/appStyle'
 import { t } from 'i18next'
+import useTimer from './utils/hooks/useTimer.jsx'
+import usePinValidation from './utils/hooks/usePinValidation.jsx'
 
 const PinFormComponent = ({
   name,
@@ -22,56 +24,15 @@ const PinFormComponent = ({
   validatePin,
   fieldHint,
 }) => {
-  const [pin, setPin] = useState(new Array(length).fill(''))
-  const [timeLeft, setTimeLeft] = useState(timerConfig.duration)
-  const [isTimerActive, setIsTimerActive] = useState(true)
-  const [errorMessage, setErrorMessage] = useState('')
-  const [isPinValid, setIsPinValid] = useState(false)
-
-  useEffect(() => {
-    if (isTimerActive && timeLeft > 0) {
-      const timer = setTimeout(() => setTimeLeft(timeLeft - 1), 1000)
-      return () => clearTimeout(timer)
-    } else if (timeLeft === 0) {
-      setIsTimerActive(false)
-    }
-  }, [timeLeft, isTimerActive])
-
-  const handlePinChange = async (element, index) => {
-    if (isNaN(element.value)) return false
-
-    const newPin = [...pin.map((d, idx) => (idx === index ? element.value : d))]
-    setPin(newPin)
-    handleChange({ target: { name, value: newPin.join('') } })
-
-    if (element.nextSibling) {
-      element.nextSibling.focus()
-    }
-
-    if (newPin.every((digit) => digit !== '')) {
-      const isValid = await validatePin(newPin.join(''))
-      if (!isValid) {
-        setErrorMessage('Invalid PIN, please try again.')
-        setIsPinValid(false)
-        setPin(new Array(length).fill(''))
-      } else {
-        setErrorMessage('')
-        setIsPinValid(true)
-        document.activeElement.blur()
-      }
-    }
-  }
-
-  const handlePaste = (e) => {
-    const paste = e.clipboardData.getData('text')
-    const pinArray = paste.split('').slice(0, length)
-    setPin(pinArray)
-    handleChange({ target: { name, value: pinArray.join('') } })
-  }
+  const { pin, errorMessage, isPinValid, handlePinChange, handlePaste } =
+    usePinValidation(length, validatePin, handleChange)
+  const [timeLeft, isTimerActive, resetTimer] = useTimer(
+    timerConfig.duration,
+    true
+  )
 
   const handleResend = () => {
-    setTimeLeft(timerConfig.duration)
-    setIsTimerActive(true)
+    resetTimer(timerConfig.duration)
   }
 
   return (
@@ -112,17 +73,7 @@ const PinFormComponent = ({
       >
         {fieldHint}
       </FormHelperText>
-      {errorMessage && (
-        <Text
-          color="layout.darkRed.darkRed400"
-          textAlign="center"
-          mt={2}
-          w="100%"
-        >
-          {errorMessage}
-        </Text>
-      )}
-      <HStack spacing={2} justify="end" mt={4}>
+      <HStack spacing={2} justify="end" margin="0 54px 15px">
         {isTimerActive && (
           <Text
             color={!darkTheme ? 'layout.white.white0' : 'layout.black.black850'}
@@ -139,6 +90,16 @@ const PinFormComponent = ({
           {timerConfig.resendLabel}
         </Button>
       </HStack>
+      {errorMessage && (
+        <Text
+          color="layout.darkRed.darkRed400"
+          textAlign="center"
+          mt={2}
+          w="100%"
+        >
+          {errorMessage}
+        </Text>
+      )}
     </Container>
   )
 }
