+++ src/ui/components/Form/utils/hooks/useTimer.jsx
@@ -1,0 +1,24 @@
+import { useState, useEffect } from 'react'
+
+const useTimer = (initialTime, isActive) => {
+  const [timeLeft, setTimeLeft] = useState(initialTime)
+  const [isTimerActive, setIsTimerActive] = useState(isActive)
+
+  useEffect(() => {
+    if (isTimerActive && timeLeft > 0) {
+      const timer = setTimeout(() => setTimeLeft(timeLeft - 1), 1000)
+      return () => clearTimeout(timer)
+    } else if (timeLeft === 0) {
+      setIsTimerActive(false)
+    }
+  }, [timeLeft, isTimerActive])
+
+  const resetTimer = (newTime) => {
+    setTimeLeft(newTime)
+    setIsTimerActive(true)
+  }
+
+  return [timeLeft, isTimerActive, resetTimer]
+}
+
+export default useTimer
